新增 shop service 初步功能以及實現

+++ models/stock.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type Stock struct {
-	ID               int       `json:"id"`
+	ID               uint64    `json:"id"`
 	ProductID        string    `json:"product_id"`
-	Quantity         int       `json:"quantity"`
-	ReservedQuantity int       `json:"reserved_quantity"`
+	Quantity         uint64    `json:"quantity"`
+	ReservedQuantity uint64    `json:"reserved_quantity"`
 	Location         string    `json:"location"`
 	CreatedAt        time.Time `json:"created_at"`
 	UpdatedAt        time.Time `json:"updated_at"`
